# samples/server/petstore/springboot/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: "This spec is mainly for testing Petstore server and contains fake\
    \ endpoints, models. Please do not use this for any other purpose. Special characters:\
    \ \" \\"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAPI Petstore
  version: 1.0.0
servers:
- url: http://petstore.swagger.io:80/v2
...
paths:
  /issue:
    post:
      operationId: reproduceIssue
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/reproduceIssue_request'
        required: false
      responses:
        "200":
          content: {}
          description: successful operation
      tags:
      - pet
      x-codegen-request-body-name: body
      x-content-type: application/x-www-form-urlencoded
      x-accepts: application/json
      x-tags:
      - tag: pet
...
components:
  schemas:
...
    reproduceIssue_request:
      properties:
        listOfStrings:
          items:
            type: string
          type: array
      type: object
...
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/api/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey
    api_key_query:
      in: query
      name: api_key_query
      type: apiKey
    http_basic_test:
      scheme: basic
      type: http
...